requirements.txt
pillow
pandas
openpyxl
xlsxwriter
pyinstaller
